+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 # Todo
 - none
 
-# v2024.6.1-beta
+# v2024.6.25
 - jslint - Remove unnecessary shell-function shCurlExe().
 - lgbm - Upgrade to lightgbm-v4.4.0.
 - jslint - Upgrade to jslint-v2024.6.1-beta.